requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
pdf2image==1.17.0
pillow==11.0.0
PyPDF2==3.0.1
pytesseract==0.3.13
Werkzeug==3.0.4
